issue40 fixed.
