+++ SlicerCART/src/configuration_config.yml
@@ -41,7 +41,6 @@ checkboxes:
   SAH_checkbox: SAH
   SDH_checkbox: SDH
   Tonsillar_herniation_checkbox: Tonsillar herniation
-
 comboboxes:
   atrophy:
     atrophy_mild: Mild
@@ -92,13 +91,5 @@ labels:
   name: IVH
   upper_bound_HU: 90
   value: 2
-- color_b: 255
-  color_g: 80
-  color_r: 11
-  lower_bound_HU: 5
-  name: PHE
-  upper_bound_HU: 33
-  value: 3
-
 modality: MRI
 slice_view_color: Yellow
