+++ SlicerCART/src/scripts/OptionalMethods.py
@@ -1,0 +1,4 @@
+from utils import *
+
+class OptionalMethods():
+    pass
