+++ SlicerCART/src/scripts/Timer.py
@@ -1,0 +1,23 @@
+from utils import *
+class Timer():
+    def __init__(self, number=None):
+        with TIMER_MUTEX:
+            self.number = number
+            self.total_time = 0
+            self.inter_time = 0
+            # counting flag to allow to PAUSE the time
+            self.flag = False  # False = not counting, True = counting (for pause button)
+
+    def start(self):
+        with TIMER_MUTEX:
+            if self.flag == False:
+                # start counting flag (to allow to pause the time if False)
+                self.flag = True
+                self.start_time = time.time()
+
+    def stop(self):
+        with TIMER_MUTEX:
+            if self.flag == True:
+                self.inter_time = time.time() - self.start_time
+                self.total_time += self.inter_time
+                self.flag = False
