+++ SlicerCART/src/scripts/__init__.py
@@ -1,1 +1,10 @@
-from .megafile import *
+from .InteractingClasses import *
+from .CompareSegmentVersionsWindow import *
+from .CustomInteractorStyle import *
+from .LoadClassificationWindow import *
+from .LoadSegmentationWindow import *
+from .OptionalMethods import *
+from .ShowSegmentVersionLegendWindow import *
+from .SlicerCARTLogic import *
+from .SlicerCARTTest import *
+from .Timer import *
