+++ SlicerCART/src/utils/__init__.py
@@ -1,4 +1,3 @@
 from .global_variables import *
 from .install_python_packages import *
 from .requirements import *
-from .functions_to_import import *
